Extract MockDelegateBridge class (#4192)
* Extract MockDelegateBridge class

All methods in GameDataTestUtil related to mocking instances of
IDelegateBrdige were moved to the new class to improve cohesion.

* Rename getDelegateBridge() -> newDelegateBridge()
