+++ game-core/src/test/java/games/strategy/triplea/delegate/GameDataTestUtil.java
@@ -1,27 +1,13 @@
 package games.strategy.triplea.delegate;
 
 import static com.google.common.base.Preconditions.checkNotNull;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyInt;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.doAnswer;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 
-import java.net.InetAddress;
-import java.net.UnknownHostException;
 import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
 
 import org.junit.jupiter.api.Assertions;
-import org.mockito.stubbing.Answer;
-import org.mockito.stubbing.OngoingStubbing;
-import org.mockito.verification.VerificationMode;
 
-import games.strategy.engine.data.Change;
 import games.strategy.engine.data.GameData;
 import games.strategy.engine.data.PlayerID;
 import games.strategy.engine.data.Route;
@@ -32,20 +18,8 @@ import games.strategy.engine.data.UnitType;
 import games.strategy.engine.data.changefactory.ChangeFactory;
 import games.strategy.engine.data.properties.BooleanProperty;
 import games.strategy.engine.data.properties.IEditableProperty;
-import games.strategy.engine.delegate.IDelegateBridge;
-import games.strategy.engine.history.DelegateHistoryWriter;
-import games.strategy.engine.history.History;
-import games.strategy.engine.history.HistoryWriter;
-import games.strategy.engine.history.IDelegateHistoryWriter;
-import games.strategy.engine.message.ChannelMessenger;
-import games.strategy.engine.message.unifiedmessenger.UnifiedMessenger;
-import games.strategy.net.IServerMessenger;
-import games.strategy.net.Node;
-import games.strategy.sound.ISound;
 import games.strategy.triplea.Constants;
 import games.strategy.triplea.attachments.TechAttachment;
-import games.strategy.triplea.player.ITripleAPlayer;
-import games.strategy.triplea.ui.display.ITripleADisplay;
 import games.strategy.util.IntegerMap;
 import junit.framework.AssertionFailedError;
 
@@ -411,83 +385,6 @@ public final class GameDataTestUtil {
    */
   static BidPlaceDelegate bidPlaceDelegate(final GameData data) {
     return (BidPlaceDelegate) data.getDelegateList().getDelegate("placeBid");
-  }
-
-  /**
-   * Returns a delegate bridge suitable for testing the given player and game data.
-   *
-   * @return A mock that can be configured using standard Mockito idioms.
-   */
-  static IDelegateBridge getDelegateBridge(final PlayerID playerId, final GameData gameData) {
-    final IDelegateBridge delegateBridge = mock(IDelegateBridge.class);
-    doAnswer(invocation -> {
-      final Change change = invocation.getArgument(0);
-      gameData.performChange(change);
-      return null;
-    }).when(delegateBridge).addChange(any());
-    when(delegateBridge.getData()).thenReturn(gameData);
-    when(delegateBridge.getDisplayChannelBroadcaster()).thenReturn(mock(ITripleADisplay.class));
-    final IDelegateHistoryWriter delegateHistoryWriter = newFakeDelegateHistoryWriter(gameData);
-    when(delegateBridge.getHistoryWriter()).thenReturn(delegateHistoryWriter);
-    when(delegateBridge.getPlayerId()).thenReturn(playerId);
-    final ITripleAPlayer remotePlayer = mock(ITripleAPlayer.class);
-    when(delegateBridge.getRemotePlayer()).thenReturn(remotePlayer);
-    when(delegateBridge.getRemotePlayer(any())).thenReturn(remotePlayer);
-    when(delegateBridge.getSoundChannelBroadcaster()).thenReturn(mock(ISound.class));
-    return delegateBridge;
-  }
-
-  private static IDelegateHistoryWriter newFakeDelegateHistoryWriter(final GameData gameData) {
-    final HistoryWriter historyWriter = new HistoryWriter(new History(gameData));
-    historyWriter.startNextStep("", "", PlayerID.NULL_PLAYERID, "");
-    final IServerMessenger serverMessenger = mock(IServerMessenger.class);
-    try {
-      when(serverMessenger.getLocalNode()).thenReturn(new Node("dummy", InetAddress.getLocalHost(), 0));
-    } catch (final UnknownHostException e) {
-      throw new IllegalStateException("test cannot run without network interface", e);
-    }
-    when(serverMessenger.isServer()).thenReturn(true);
-    return new DelegateHistoryWriter(new ChannelMessenger(new UnifiedMessenger(serverMessenger)));
-  }
-
-  static OngoingStubbing<int[]> whenGetRandom(final IDelegateBridge delegateBridge) {
-    return when(delegateBridge.getRandom(anyInt(), anyInt(), any(), any(), anyString()));
-  }
-
-  static Answer<int[]> withValues(final int... values) {
-    return invocation -> {
-      final int count = invocation.getArgument(1);
-      assertEquals(values.length, count, "count of requested random values does not match");
-      return values;
-    };
-  }
-
-  static void thenGetRandomShouldHaveBeenCalled(
-      final IDelegateBridge delegateBridge,
-      final VerificationMode verificationMode) {
-    verify(delegateBridge, verificationMode).getRandom(anyInt(), anyInt(), any(), any(), anyString());
-  }
-
-  static ITripleAPlayer withRemotePlayer(final IDelegateBridge delegateBridge) {
-    return (ITripleAPlayer) delegateBridge.getRemotePlayer();
-  }
-
-  static void advanceToStep(final IDelegateBridge delegateBridge, final String stepName) {
-    final GameData gameData = delegateBridge.getData();
-    gameData.acquireWriteLock();
-    try {
-      final int length = gameData.getSequence().size();
-      int i = 0;
-      while ((i < length) && (gameData.getSequence().getStep().getName().indexOf(stepName) == -1)) {
-        gameData.getSequence().next();
-        i++;
-      }
-      if ((i > length) && (gameData.getSequence().getStep().getName().indexOf(stepName) == -1)) {
-        throw new IllegalArgumentException("Step not found: " + stepName);
-      }
-    } finally {
-      gameData.releaseWriteLock();
-    }
   }
 
   static void load(final Collection<Unit> units, final Route route) {
