+++ game-core/src/test/java/games/strategy/triplea/delegate/RevisedTest.java
@@ -10,7 +10,6 @@ import static games.strategy.triplea.delegate.BattleStepStrings.SUBS_FIRE;
 import static games.strategy.triplea.delegate.BattleStepStrings.SUBS_SUBMERGE;
 import static games.strategy.triplea.delegate.GameDataTestUtil.aaGun;
 import static games.strategy.triplea.delegate.GameDataTestUtil.addTo;
-import static games.strategy.triplea.delegate.GameDataTestUtil.advanceToStep;
 import static games.strategy.triplea.delegate.GameDataTestUtil.americans;
 import static games.strategy.triplea.delegate.GameDataTestUtil.armour;
 import static games.strategy.triplea.delegate.GameDataTestUtil.assertError;
@@ -34,11 +33,12 @@ import static games.strategy.triplea.delegate.GameDataTestUtil.removeFrom;
 import static games.strategy.triplea.delegate.GameDataTestUtil.submarine;
 import static games.strategy.triplea.delegate.GameDataTestUtil.techDelegate;
 import static games.strategy.triplea.delegate.GameDataTestUtil.territory;
-import static games.strategy.triplea.delegate.GameDataTestUtil.thenGetRandomShouldHaveBeenCalled;
 import static games.strategy.triplea.delegate.GameDataTestUtil.transport;
-import static games.strategy.triplea.delegate.GameDataTestUtil.whenGetRandom;
-import static games.strategy.triplea.delegate.GameDataTestUtil.withRemotePlayer;
-import static games.strategy.triplea.delegate.GameDataTestUtil.withValues;
+import static games.strategy.triplea.delegate.MockDelegateBridge.advanceToStep;
+import static games.strategy.triplea.delegate.MockDelegateBridge.thenGetRandomShouldHaveBeenCalled;
+import static games.strategy.triplea.delegate.MockDelegateBridge.whenGetRandom;
+import static games.strategy.triplea.delegate.MockDelegateBridge.withRemotePlayer;
+import static games.strategy.triplea.delegate.MockDelegateBridge.withValues;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -122,8 +122,8 @@ public class RevisedTest {
     gameData = TestMapGameData.REVISED.getGameData();
   }
 
-  private IDelegateBridge getDelegateBridge(final PlayerID player) {
-    return GameDataTestUtil.getDelegateBridge(player, gameData);
+  private IDelegateBridge newDelegateBridge(final PlayerID player) {
+    return MockDelegateBridge.newInstance(gameData, player);
   }
 
   @Test
@@ -132,7 +132,7 @@ public class RevisedTest {
     final Territory sz1 = gameData.getMap().getTerritory("1 Sea Zone");
     final Territory sz11 = gameData.getMap().getTerritory("11 Sea Zone");
     final Territory sz9 = gameData.getMap().getTerritory("9 Sea Zone");
-    final IDelegateBridge bridge = getDelegateBridge(british);
+    final IDelegateBridge bridge = newDelegateBridge(british);
     advanceToStep(bridge, "NonCombatMove");
     moveDelegate(gameData).setDelegateBridgeAndPlayer(bridge);
     moveDelegate(gameData).start();
@@ -157,7 +157,7 @@ public class RevisedTest {
     // before the advance, subs defend and attack at 2
     assertEquals(2, attachment.getDefense(japanese));
     assertEquals(2, attachment.getAttack(japanese));
-    final IDelegateBridge bridge = getDelegateBridge(japanese);
+    final IDelegateBridge bridge = newDelegateBridge(japanese);
     TechTracker.addAdvance(japanese, bridge,
         TechAdvance.findAdvance(TechAdvance.TECH_PROPERTY_SUPER_SUBS, gameData, japanese));
     // after tech advance, this is now 3
@@ -175,7 +175,7 @@ public class RevisedTest {
     sub.setSubmerged(true);
     // now move to attack it
     final MoveDelegate moveDelegate = (MoveDelegate) gameData.getDelegateList().getDelegate("move");
-    final IDelegateBridge bridge = getDelegateBridge(british);
+    final IDelegateBridge bridge = newDelegateBridge(british);
     advanceToStep(bridge, "NonCombatMove");
     moveDelegate.setDelegateBridgeAndPlayer(bridge);
     moveDelegate.start();
@@ -193,7 +193,7 @@ public class RevisedTest {
   public void testRetreatBug() {
     final PlayerID russians = GameDataTestUtil.russians(gameData);
     final PlayerID americans = GameDataTestUtil.americans(gameData);
-    final IDelegateBridge bridge = getDelegateBridge(russians);
+    final IDelegateBridge bridge = newDelegateBridge(russians);
     // we need to initialize the original owner
     final InitializationDelegate initDel =
         (InitializationDelegate) gameData.getDelegateList().getDelegate("initDelegate");
@@ -246,7 +246,7 @@ public class RevisedTest {
   public void testContinuedBattles() {
     final PlayerID russians = GameDataTestUtil.russians(gameData);
     final PlayerID germans = germans(gameData);
-    final IDelegateBridge bridge = getDelegateBridge(germans);
+    final IDelegateBridge bridge = newDelegateBridge(germans);
     final MoveDelegate moveDelegate = (MoveDelegate) gameData.getDelegateList().getDelegate("move");
     advanceToStep(bridge, "CombatMove");
     moveDelegate.setDelegateBridgeAndPlayer(bridge);
@@ -292,7 +292,7 @@ public class RevisedTest {
     final PlayerID british = british(gameData);
     final PlayerID americans = americans(gameData);
     final Territory uk = territory("United Kingdom", gameData);
-    final IDelegateBridge bridge = getDelegateBridge(british);
+    final IDelegateBridge bridge = newDelegateBridge(british);
     advanceToStep(bridge, "CombatMove");
     moveDelegate(gameData).setDelegateBridgeAndPlayer(bridge);
     moveDelegate(gameData).start();
@@ -308,7 +308,7 @@ public class RevisedTest {
 
   @Test
   public void testBidPlace() {
-    final IDelegateBridge bridge = getDelegateBridge(british(gameData));
+    final IDelegateBridge bridge = newDelegateBridge(british(gameData));
     advanceToStep(bridge, "BidPlace");
     bidPlaceDelegate(gameData).setDelegateBridgeAndPlayer(bridge);
     bidPlaceDelegate(gameData).start();
@@ -327,7 +327,7 @@ public class RevisedTest {
   public void testOverFlyBombersDies() {
     final PlayerID british = british(gameData);
     final MoveDelegate moveDelegate = (MoveDelegate) gameData.getDelegateList().getDelegate("move");
-    final IDelegateBridge bridge = getDelegateBridge(british);
+    final IDelegateBridge bridge = newDelegateBridge(british);
     advanceToStep(bridge, "CombatMove");
     moveDelegate.setDelegateBridgeAndPlayer(bridge);
     moveDelegate.start();
@@ -348,7 +348,7 @@ public class RevisedTest {
   public void testMultipleOverFlyBombersDies() {
     final PlayerID british = british(gameData);
     final MoveDelegate moveDelegate = (MoveDelegate) gameData.getDelegateList().getDelegate("move");
-    final IDelegateBridge bridge = getDelegateBridge(british);
+    final IDelegateBridge bridge = newDelegateBridge(british);
     advanceToStep(bridge, "CombatMove");
     moveDelegate.setDelegateBridgeAndPlayer(bridge);
     moveDelegate.start();
@@ -377,7 +377,7 @@ public class RevisedTest {
     // it should not appear in the battle
     final PlayerID british = british(gameData);
     final MoveDelegate moveDelegate = (MoveDelegate) gameData.getDelegateList().getDelegate("move");
-    final IDelegateBridge bridge = getDelegateBridge(british);
+    final IDelegateBridge bridge = newDelegateBridge(british);
     advanceToStep(bridge, "CombatMove");
     moveDelegate.setDelegateBridgeAndPlayer(bridge);
     moveDelegate.start();
@@ -400,7 +400,7 @@ public class RevisedTest {
     final Territory sz13 = gameData.getMap().getTerritory("13 Sea Zone");
     final PlayerID germans = GameDataTestUtil.germans(gameData);
     final MoveDelegate moveDelegate = (MoveDelegate) gameData.getDelegateList().getDelegate("move");
-    final IDelegateBridge bridge = getDelegateBridge(germans);
+    final IDelegateBridge bridge = newDelegateBridge(germans);
     advanceToStep(bridge, "CombatMove");
     moveDelegate.setDelegateBridgeAndPlayer(bridge);
     moveDelegate.start();
@@ -420,7 +420,7 @@ public class RevisedTest {
     final UnitType infantryType = GameDataTestUtil.infantry(gameData);
     final PlayerID germans = GameDataTestUtil.germans(gameData);
     final MoveDelegate moveDelegate = (MoveDelegate) gameData.getDelegateList().getDelegate("move");
-    final IDelegateBridge bridge = getDelegateBridge(germans);
+    final IDelegateBridge bridge = newDelegateBridge(germans);
     advanceToStep(bridge, "CombatMove");
     moveDelegate.setDelegateBridgeAndPlayer(bridge);
     moveDelegate.start();
@@ -453,7 +453,7 @@ public class RevisedTest {
     final UnitType infantryType = GameDataTestUtil.infantry(gameData);
     final PlayerID germans = GameDataTestUtil.germans(gameData);
     final MoveDelegate moveDelegate = (MoveDelegate) gameData.getDelegateList().getDelegate("move");
-    final IDelegateBridge bridge = getDelegateBridge(germans);
+    final IDelegateBridge bridge = newDelegateBridge(germans);
     advanceToStep(bridge, "CombatMove");
     moveDelegate.setDelegateBridgeAndPlayer(bridge);
     moveDelegate.start();
@@ -499,7 +499,7 @@ public class RevisedTest {
     final UnitType infantryType = GameDataTestUtil.infantry(gameData);
     final PlayerID germans = GameDataTestUtil.germans(gameData);
     final MoveDelegate moveDelegate = (MoveDelegate) gameData.getDelegateList().getDelegate("move");
-    final IDelegateBridge bridge = getDelegateBridge(germans);
+    final IDelegateBridge bridge = newDelegateBridge(germans);
     advanceToStep(bridge, "CombatMove");
     moveDelegate.setDelegateBridgeAndPlayer(bridge);
     moveDelegate.start();
@@ -538,7 +538,7 @@ public class RevisedTest {
     gameData.performChange(change);
     final Territory sz5 = gameData.getMap().getTerritory("5 Sea Zone");
     final MoveDelegate moveDelegate = (MoveDelegate) gameData.getDelegateList().getDelegate("move");
-    final IDelegateBridge bridge = getDelegateBridge(japanese);
+    final IDelegateBridge bridge = newDelegateBridge(japanese);
     advanceToStep(bridge, "CombatMove");
     moveDelegate.setDelegateBridgeAndPlayer(bridge);
     moveDelegate.start();
@@ -568,7 +568,7 @@ public class RevisedTest {
     final UnitType infantryType = GameDataTestUtil.infantry(gameData);
     final PlayerID germans = GameDataTestUtil.germans(gameData);
     final MoveDelegate moveDelegate = (MoveDelegate) gameData.getDelegateList().getDelegate("move");
-    final IDelegateBridge bridge = getDelegateBridge(germans);
+    final IDelegateBridge bridge = newDelegateBridge(germans);
     advanceToStep(bridge, "CombatMove");
     moveDelegate.setDelegateBridgeAndPlayer(bridge);
     moveDelegate.start();
@@ -621,7 +621,7 @@ public class RevisedTest {
     final UnitType infantryType = GameDataTestUtil.infantry(gameData);
     final PlayerID germans = GameDataTestUtil.germans(gameData);
     final MoveDelegate moveDelegate = (MoveDelegate) gameData.getDelegateList().getDelegate("move");
-    final IDelegateBridge bridge = getDelegateBridge(germans);
+    final IDelegateBridge bridge = newDelegateBridge(germans);
     advanceToStep(bridge, "CombatMove");
     moveDelegate.setDelegateBridgeAndPlayer(bridge);
     moveDelegate.start();
@@ -665,7 +665,7 @@ public class RevisedTest {
     // german sub tries to attack a transport in non combat
     // should be an error
     final MoveDelegate moveDelegate = (MoveDelegate) gameData.getDelegateList().getDelegate("move");
-    final IDelegateBridge bridge = getDelegateBridge(germans);
+    final IDelegateBridge bridge = newDelegateBridge(germans);
     advanceToStep(bridge, "NonCombatMove");
     moveDelegate.setDelegateBridgeAndPlayer(bridge);
     moveDelegate.start();
@@ -681,7 +681,7 @@ public class RevisedTest {
     // german sub tries to attack a transport in non combat
     // should be an error
     final MoveDelegate moveDelegate = (MoveDelegate) gameData.getDelegateList().getDelegate("move");
-    final IDelegateBridge bridge = getDelegateBridge(germans);
+    final IDelegateBridge bridge = newDelegateBridge(germans);
     advanceToStep(bridge, "NonCombatMove");
     moveDelegate.setDelegateBridgeAndPlayer(bridge);
     moveDelegate.start();
@@ -697,7 +697,7 @@ public class RevisedTest {
     // german sub tries to attack a transport in non combat
     // should be an error
     final MoveDelegate moveDelegate = (MoveDelegate) gameData.getDelegateList().getDelegate("move");
-    final IDelegateBridge bridge = getDelegateBridge(british);
+    final IDelegateBridge bridge = newDelegateBridge(british);
     advanceToStep(bridge, "NonCombatMove");
     moveDelegate.setDelegateBridgeAndPlayer(bridge);
     moveDelegate.start();
@@ -717,7 +717,7 @@ public class RevisedTest {
     gameData.performChange(c);
     // new move delegate
     final MoveDelegate moveDelegate = (MoveDelegate) gameData.getDelegateList().getDelegate("move");
-    final IDelegateBridge bridge = getDelegateBridge(japanese);
+    final IDelegateBridge bridge = newDelegateBridge(japanese);
     advanceToStep(bridge, "CombatMove");
     moveDelegate.setDelegateBridgeAndPlayer(bridge);
     moveDelegate.start();
@@ -775,7 +775,7 @@ public class RevisedTest {
     final String preOwner = fic.getOwner().getName();
     assertEquals(Constants.PLAYER_NAME_JAPANESE, preOwner);
     // Set up the move delegate
-    IDelegateBridge delegateBridge = getDelegateBridge(british(gameData));
+    IDelegateBridge delegateBridge = newDelegateBridge(british(gameData));
     final MoveDelegate moveDelegate = moveDelegate(gameData);
     advanceToStep(delegateBridge, "CombatMove");
     moveDelegate.setDelegateBridgeAndPlayer(delegateBridge);
@@ -802,7 +802,7 @@ public class RevisedTest {
      * add a VALID JAPANESE attack
      */
     // Set up battle
-    delegateBridge = getDelegateBridge(japanese(gameData));
+    delegateBridge = newDelegateBridge(japanese(gameData));
     advanceToStep(delegateBridge, "CombatMove");
     moveDelegate.setDelegateBridgeAndPlayer(delegateBridge);
     moveDelegate.start();
@@ -824,7 +824,7 @@ public class RevisedTest {
      * add a VALID AMERICAN attack
      */
     // Set up battle
-    delegateBridge = getDelegateBridge(americans(gameData));
+    delegateBridge = newDelegateBridge(americans(gameData));
     advanceToStep(delegateBridge, "CombatMove");
     moveDelegate.setDelegateBridgeAndPlayer(delegateBridge);
     moveDelegate.start();
@@ -856,7 +856,7 @@ public class RevisedTest {
     addTo(germany, bombers);
     battle.addAttackChange(gameData.getMap().getRoute(uk, germany), bombers, null);
     tracker.getBattleRecords().addBattle(british, battle.getBattleId(), germany, battle.getBattleType());
-    final IDelegateBridge bridge = getDelegateBridge(british);
+    final IDelegateBridge bridge = newDelegateBridge(british);
     // aa guns rolls 0 and hits
     whenGetRandom(bridge).thenAnswer(withValues(0));
     final int pusBeforeRaid = germans.getResources().getQuantity(gameData.getResourceList().getResource(Constants.PUS));
@@ -880,7 +880,7 @@ public class RevisedTest {
     addTo(germany, bombers);
     battle.addAttackChange(gameData.getMap().getRoute(uk, germany), bombers, null);
     tracker.getBattleRecords().addBattle(british, battle.getBattleId(), germany, battle.getBattleType());
-    final IDelegateBridge bridge = getDelegateBridge(british);
+    final IDelegateBridge bridge = newDelegateBridge(british);
     // should be exactly 3 rolls total. would be exactly 2 rolls if the number of units being shot at = max dice side of
     // the AA gun, because
     // the casualty selection roll would not happen in LL
@@ -912,7 +912,7 @@ public class RevisedTest {
     addTo(germany, bombers);
     battle.addAttackChange(gameData.getMap().getRoute(uk, germany), bombers, null);
     tracker.getBattleRecords().addBattle(british, battle.getBattleId(), germany, battle.getBattleType());
-    final IDelegateBridge bridge = getDelegateBridge(british);
+    final IDelegateBridge bridge = newDelegateBridge(british);
     // aa guns rolls 0 and hits, next 5 dice are for the bombing raid cost for the
     // surviving bombers
     whenGetRandom(bridge)
@@ -939,7 +939,7 @@ public class RevisedTest {
         uk.getUnits().getMatches(Matches.unitIsStrategicBomber()), null);
     addTo(germany, uk.getUnits().getMatches(Matches.unitIsStrategicBomber()));
     tracker.getBattleRecords().addBattle(british, battle.getBattleId(), germany, battle.getBattleType());
-    final IDelegateBridge bridge = getDelegateBridge(british);
+    final IDelegateBridge bridge = newDelegateBridge(british);
     TechTracker.addAdvance(british, bridge,
         TechAdvance.findAdvance(TechAdvance.TECH_PROPERTY_HEAVY_BOMBER, gameData, british));
     // aa guns rolls 3, misses, bomber rolls 2 dice at 3
@@ -958,7 +958,7 @@ public class RevisedTest {
     final String attacker = "British";
     final Territory attacked = territory("Libya", gameData);
     final Territory from = territory("Anglo Egypt", gameData);
-    final IDelegateBridge bridge = getDelegateBridge(british(gameData));
+    final IDelegateBridge bridge = newDelegateBridge(british(gameData));
     advanceToStep(bridge, "CombatMove");
     moveDelegate(gameData).setDelegateBridgeAndPlayer(bridge);
     moveDelegate(gameData).start();
@@ -980,7 +980,7 @@ public class RevisedTest {
     // 1 destroyer attacks 1 destroyer
     addTo(from, destroyer(gameData).create(1, british(gameData)));
     addTo(attacked, destroyer(gameData).create(1, germans(gameData)));
-    final IDelegateBridge bridge = getDelegateBridge(british(gameData));
+    final IDelegateBridge bridge = newDelegateBridge(british(gameData));
     advanceToStep(bridge, "CombatMove");
     moveDelegate(gameData).setDelegateBridgeAndPlayer(bridge);
     moveDelegate(gameData).start();
@@ -1002,7 +1002,7 @@ public class RevisedTest {
     // 1 sub attacks 1 sub
     addTo(from, submarine(gameData).create(1, british(gameData)));
     addTo(attacked, submarine(gameData).create(1, germans(gameData)));
-    final IDelegateBridge bridge = getDelegateBridge(british(gameData));
+    final IDelegateBridge bridge = newDelegateBridge(british(gameData));
     advanceToStep(bridge, "CombatMove");
     moveDelegate(gameData).setDelegateBridgeAndPlayer(bridge);
     moveDelegate(gameData).start();
@@ -1040,7 +1040,7 @@ public class RevisedTest {
     addTo(from, submarine(gameData).create(2, british(gameData)));
     addTo(attacked, submarine(gameData).create(1, germans(gameData)));
     addTo(attacked, destroyer(gameData).create(1, germans(gameData)));
-    final IDelegateBridge bridge = getDelegateBridge(british(gameData));
+    final IDelegateBridge bridge = newDelegateBridge(british(gameData));
     advanceToStep(bridge, "CombatMove");
     moveDelegate(gameData).setDelegateBridgeAndPlayer(bridge);
     moveDelegate(gameData).start();
@@ -1099,7 +1099,7 @@ public class RevisedTest {
     addTo(from, battleship(gameData).create(1, british(gameData)));
     addTo(attacked, submarine(gameData).create(3, germans(gameData)));
     addTo(attacked, destroyer(gameData).create(1, germans(gameData)));
-    final IDelegateBridge bridge = getDelegateBridge(british(gameData));
+    final IDelegateBridge bridge = newDelegateBridge(british(gameData));
     advanceToStep(bridge, "CombatMove");
     moveDelegate(gameData).setDelegateBridgeAndPlayer(bridge);
     moveDelegate(gameData).start();
@@ -1160,7 +1160,7 @@ public class RevisedTest {
     addTo(from, submarine(gameData).create(1, british(gameData)));
     addTo(from, destroyer(gameData).create(1, british(gameData)));
     addTo(attacked, submarine(gameData).create(1, germans(gameData)));
-    final IDelegateBridge bridge = getDelegateBridge(british(gameData));
+    final IDelegateBridge bridge = newDelegateBridge(british(gameData));
     advanceToStep(bridge, "CombatMove");
     moveDelegate(gameData).setDelegateBridgeAndPlayer(bridge);
     moveDelegate(gameData).start();
@@ -1200,7 +1200,7 @@ public class RevisedTest {
     addTo(from, destroyer(gameData).create(1, british(gameData)));
     addTo(attacked, submarine(gameData).create(1, germans(gameData)));
     addTo(attacked, battleship(gameData).create(1, germans(gameData)));
-    final IDelegateBridge bridge = getDelegateBridge(british(gameData));
+    final IDelegateBridge bridge = newDelegateBridge(british(gameData));
     advanceToStep(bridge, "CombatMove");
     moveDelegate(gameData).setDelegateBridgeAndPlayer(bridge);
     moveDelegate(gameData).start();
@@ -1262,7 +1262,7 @@ public class RevisedTest {
     addTo(from, destroyer(gameData).create(1, british(gameData)));
     addTo(attacked, submarine(gameData).create(1, germans(gameData)));
     addTo(attacked, destroyer(gameData).create(1, germans(gameData)));
-    final IDelegateBridge bridge = getDelegateBridge(british(gameData));
+    final IDelegateBridge bridge = newDelegateBridge(british(gameData));
     advanceToStep(bridge, "CombatMove");
     moveDelegate(gameData).setDelegateBridgeAndPlayer(bridge);
     moveDelegate(gameData).start();
@@ -1296,7 +1296,7 @@ public class RevisedTest {
   @Test
   public void testUnplacedDie() {
     final PlaceDelegate del = placeDelegate(gameData);
-    del.setDelegateBridgeAndPlayer(getDelegateBridge(british(gameData)));
+    del.setDelegateBridgeAndPlayer(newDelegateBridge(british(gameData)));
     del.start();
     addTo(british(gameData), transport(gameData).create(1, british(gameData)), gameData);
     del.end();
@@ -1307,7 +1307,7 @@ public class RevisedTest {
   @Test
   public void testRocketsDontFireInConquered() {
     final MoveDelegate move = moveDelegate(gameData);
-    final IDelegateBridge bridge = getDelegateBridge(germans(gameData));
+    final IDelegateBridge bridge = newDelegateBridge(germans(gameData));
     advanceToStep(bridge, "CombatMove");
     move.setDelegateBridgeAndPlayer(bridge);
     move.start();
@@ -1326,7 +1326,7 @@ public class RevisedTest {
   public void testTechRolls() {
     // Set up the test
     final PlayerID germans = GameDataTestUtil.germans(gameData);
-    final IDelegateBridge delegateBridge = getDelegateBridge(germans);
+    final IDelegateBridge delegateBridge = newDelegateBridge(germans);
     advanceToStep(delegateBridge, "germanTech");
     final TechnologyDelegate techDelegate = techDelegate(gameData);
     techDelegate.setDelegateBridgeAndPlayer(delegateBridge);
@@ -1382,7 +1382,7 @@ public class RevisedTest {
     addTo(sz6, destroyer(gameData).create(2, british));
     addTo(sz5, transport(gameData).create(3, germans));
     addTo(germany, armour(gameData).create(3, germans));
-    final IDelegateBridge bridge = getDelegateBridge(germans(gameData));
+    final IDelegateBridge bridge = newDelegateBridge(germans(gameData));
     advanceToStep(bridge, "CombatMove");
     givenRemotePlayerWillSelectDefaultCasualties(bridge);
     moveDelegate(gameData).setDelegateBridgeAndPlayer(bridge);
@@ -1419,7 +1419,7 @@ public class RevisedTest {
   public void testCanalMovePass() {
     final Territory sz15 = territory("15 Sea Zone", gameData);
     final Territory sz34 = territory("34 Sea Zone", gameData);
-    final IDelegateBridge bridge = getDelegateBridge(british(gameData));
+    final IDelegateBridge bridge = newDelegateBridge(british(gameData));
     advanceToStep(bridge, "CombatMove");
     final MoveDelegate moveDelegate = moveDelegate(gameData);
     moveDelegate.setDelegateBridgeAndPlayer(bridge);
@@ -1435,7 +1435,7 @@ public class RevisedTest {
     final Territory sz34 = territory("34 Sea Zone", gameData);
     // clear the british in sz 15
     removeFrom(sz15, sz15.getUnits().getUnits());
-    final IDelegateBridge bridge = getDelegateBridge(germans(gameData));
+    final IDelegateBridge bridge = newDelegateBridge(germans(gameData));
     advanceToStep(bridge, "CombatMove");
     final MoveDelegate moveDelegate = moveDelegate(gameData);
     moveDelegate.setDelegateBridgeAndPlayer(bridge);
